Display "File size is too large!" and disable editor for files > 5mb

+++ qml/CodeEditorView.qml
@@ -110,7 +110,7 @@ Item {
 				}
 				break;
 			}
-		}		
+		}
 	}
 
 	function displayOpenedDocument(index)
@@ -265,6 +265,13 @@ Item {
 			editor.displayGasEstimation(checked);
 	}
 
+	function setReadOnly(status)
+	{
+		var editor = getEditor(currentDocumentId);
+		if (editor)
+			editor.setReadOnly(status);
+	}
+
 	Component.onCompleted: projectModel.codeEditor = codeEditorView;
 
 	Connections {
@@ -314,7 +321,7 @@ Item {
 			for (var i = 0; i < editorListModel.count; i++)
 			{
 				var doc = editorListModel.get(i);
-				if (doc.path === document.path)
+				if (doc.path === document.path && !doc.readOnly)
 				{
 					fileIo.writeFile(document.path, editors.itemAt(i).item.getText());
 					break;
